requirements.txt
absl-py==2.1.0
accelerate==1.8.1
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.0
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
APScheduler==3.11.1
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.0.1
bert-score==0.3.13
blake3==1.0.4
blinker==1.4
boto3==1.35.97
botocore==1.35.98
braceexpand==0.1.7
build==1.2.2.post1
cachetools==4.2.4
cchardet==2.1.7
certifi==2021.10.8
cffi==2.0.0
charset-normalizer==2.0.12
cityhash==0.2.4.post11
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.1
contourpy==1.3.1
cramjam==2.10.0
cryptography==3.4.8
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.6.0
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.2
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.7.0
dnspython==2.7.0
docstring_parser==0.17.0
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastparquet==2024.2.0
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
func-timeout==4.3.5
gguf==0.10.0
h11==0.14.0
hf-xet==1.2.1
hiredis==2.4.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.1.0
interegular==0.3.3
ipython==8.30.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kazoo==2.10.0
keyring==23.5.0
kiwisolver==1.4.7
lark==1.2.2
latex2sympy2_extended==1.10.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lxml==4.9.4
lz4==3.1.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.8.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
more-itertools==8.10.0
mpi4py==4.1.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==13.590.44
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.0
openai==1.67.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
prettytable==2.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycares==4.11.0
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.4
pydantic_core==2.27.2
pyfiglet==1.0.4
Pygments==2.18.0
PyJWT==2.3.0
pylatexenc==2.10
pynvml==13.0.1
pyparsing==2.4.7
pyproject_hooks==1.2.0
pysmhasher==0.2.5
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-snappy==0.6.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.8
ray==2.43.0
redis==4.6.0
referencing==0.36.2
regex==2024.11.6
rich==13.9.4
rich-toolkit==0.13.2
rouge-score==0.1.2
rpds-py==0.23.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
sentencepiece==0.2.0
setuptools-scm==9.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torchao==0.11.0
torchdata==0.10.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.0
triton==3.1.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
unpaddedbase64==2.1.0
urllib3==1.26.8
uvicorn==0.34.0
uvloop==0.21.0
vertexai==1.71.1
virtualenv==20.29.3
wadllib==1.3.6
watchfiles==1.0.4
wcwidth==0.2.13
webdataset==0.2.100
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.11
xmltodict==0.12.0
xxhash==3.6.0
yarl==1.18.3
zipp==1.0.0
